requirements.txt
# Web 框架
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# 資料庫 ORM
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# 快取與佇列
redis==5.0.0
celery==5.4.0
flower==2.0.0

# 認證與安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # 兼容 passlib 的版本

# 資料驗證
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0

# 檔案處理
python-magic==0.4.27  # Linux 版本
pypdf2==3.0.0
python-docx==1.1.0
pillow==10.4.0

# 向量與 RAG
qdrant-client==1.11.0
langchain==0.3.0
langchain-openai==0.2.0
openai==1.51.0
tiktoken==0.7.0

# HTTP 客戶端
httpx==0.27.0
aiofiles==24.1.0

# 工具
python-dotenv==1.0.0

# 監控與日誌
python-json-logger==2.0.0
